requirements.txt
cymysql==0.9.1
Django==1.11.5
django-cymysql==0.5.1
olefile==0.44
Pillow==4.2.1
pytz==2017.2
